integrations/autogen_wraper.py
# cert/integrations/autogen_wrapper.py
import cert_client
from autogen import ConversableAgent

# Update your CERTInstrumentedAgent to use real LLMs
class CERTInstrumentedAgent(ConversableAgent):
    def __init__(self, llm_provider: LLMProvider, *args, **kwargs):
        super().__init__(*args, **kwargs)
        self.llm_provider = llm_provider
        self.cert_client = CERTClient()
        self.response_history = []
    
    def generate_reply(self, messages, sender=None, **kwargs):
        # Use real LLM instead of AutoGen's default
        prompt = messages[-1]["content"] if messages else ""
        response = self.llm_provider.generate(prompt)
        
        # Your existing CERT measurement logic
        self.response_history.append(response)
        if len(self.response_history) >= 3:
            # Run consistency measurement
            pass
        
        return response