requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
scipy==1.11.1
sentence-transformers==2.2.2
pydantic==2.4.2
huggingface-hub==0.16.4
transformers==4.21.0
torch>=1.9.0
python-dotenv==1.0.0
anthropic>=0.7.0
requests>=2.31.0
python-multipart>=0.0.6